Do not export balenaApi on api-binder
Instead, add `getBalenaApi` function to api-helper so other modules can
access a balena API instance.

Further reduces circular dependencies to 5

Change-type: patch
